transcription/requirements.txt
torch==2.8.0
torchaudio==2.8.0
soundfile
librosa
pyannote.audio==4.0.0
numpy==2.2.6
onnxruntime
openai-whisper
pydub
